+++ HCM_Flood_Level/Core/Entities/AlertLog.cs
@@ -8,12 +8,11 @@ namespace Core.Entities
 {
     public class AlertLog
     {
-        public int LogId { get; set; }
+        public long LogId { get; set; }
         public string Channel { get; set; }
         public DateTime SentAt { get; set; }
         public string LogStatus { get; set; }
-
-        public int AlertId { get; set; }
+        public long AlertId { get; set; }
         public Alert Alert { get; set; }
     }
 }
